refactor(demos): migrate wave 7 (display-state pair) to the shared style system
Move display-state and container-display-state flat to folder, both with zero
bespoke residue. These are the demos the badge/chip/readout/banner vocab was
seeded for, so the registry grows to serve them precisely:

- message: new shared fragment, an inline validation line that recolours
  danger on error and accent on pending. Collapses the two demos' .message and
  .msg synonyms onto one class.
- chip.on: active-projection variant on the chip fragment (the show* booleans
  light their chip).
- banner: reshaped into a tinted bordered box that works two ways, a block
  alert on its own and a flex status bar when composed with readout
  (class="readout banner"). error/failed and success/saved are grouped as
  synonyms so a display-state verdict and a save-status read share one rule.
- fieldset: the legend becomes a flex row (no change for plain text legends,
  a single child) so a group can carry its rolled-up displayState badge inline.
